requirements.txt
python_dotenv==0.19.2
coloredlogs==15.0.1
colorthief==0.2.1
psutil==5.8.0